requirements.txt
flask
requests
beautifulsoup4
pypdf
fpdf2
cloudinary
google-genai
python-dotenv
gunicorn
gevent